simple keybinding parsing
